added currency exchange in alfa version :D no tests :D

+++ AuctionSite/BL/DTOs/Auction/CreateAuction.cs
@@ -32,10 +32,10 @@ namespace BL.DTOs.Auction
         public DateTime EndDate { get; set; }
 
         [Required]
-        [Range(0, double.MaxValue, ErrorMessage = "Please enter valid start price")]
-        public double StartPrice { get; set; }
+        [Range(0, (double)decimal.MaxValue, ErrorMessage = "Please enter valid start price")]
+        public decimal StartPrice { get; set; }
 
-        public double ActualPrice { get; set; }
+        public decimal ActualPrice { get; set; }
         public IList<ImageDto> ImageBytes { get; set; } = new List<ImageDto>();
         
         public IList<HttpPostedFileBase> Upload { get; set; }
